Initial version of a format proposal
This is a *very* rough draft of an informal description of a debug
format. Several technical details have been handwaved away or likely
omitted (intentionally or not). The goal of this document is primarily
to decide if it comes close to including the information we want or
whether it serves as a good starting point for future iterations.